enforce filter dtype

+++ src/aspire/image/image.py
@@ -420,7 +420,9 @@ class Image:
         # Note image and filter data is intentionally migrated via
         # `xp.asarray` because all of the subsequent calls until
         # `asnumpy` are GPU when xp and fft in `cupy` mode.
-        filter_values = xp.asarray(filter.evaluate_grid(self.resolution))
+        filter_values = xp.asarray(
+            filter.evaluate_grid(self.resolution), dtype=self.dtype
+        )
 
         # Convolve
         im_f = fft.centered_fft2(xp.asarray(im._data))
